+++ Classes/Utilities.m
@@ -85,45 +85,45 @@
 			transform = CGAffineTransformIdentity;
 			break;
 		case UIImageOrientationUpMirrored:
-			transform = CGAffineTransformMakeTranslation(imageSize.width, 0.0);
-			transform = CGAffineTransformScale(transform, -1.0, 1.0);
+			transform = CGAffineTransformMakeTranslation(imageSize.width, (CGFloat) 0.0);
+			transform = CGAffineTransformScale(transform, (CGFloat) -1.0, (CGFloat) 1.0);
 			break;
 		case UIImageOrientationDown:
 			transform = CGAffineTransformMakeTranslation(imageSize.width, imageSize.height);
-			transform = CGAffineTransformRotate(transform, M_PI);
+			transform = CGAffineTransformRotate(transform, (CGFloat) M_PI);
 			break;
 		case UIImageOrientationDownMirrored:
-			transform = CGAffineTransformMakeTranslation(0.0, imageSize.height);
-			transform = CGAffineTransformScale(transform, 1.0, -1.0);
+			transform = CGAffineTransformMakeTranslation((CGFloat) 0.0, imageSize.height);
+			transform = CGAffineTransformScale(transform, (CGFloat) 1.0, (CGFloat) -1.0);
 			break;
 		case UIImageOrientationLeftMirrored:
 			boundHeight = bounds.size.height;
 			bounds.size.height = bounds.size.width;
 			bounds.size.width = boundHeight;
 			transform = CGAffineTransformMakeTranslation(imageSize.height, imageSize.width);
-			transform = CGAffineTransformScale(transform, -1.0, 1.0);
-			transform = CGAffineTransformRotate(transform, 3.0 * M_PI / 2.0);
+			transform = CGAffineTransformScale(transform, (CGFloat) -1.0, (CGFloat) 1.0);
+			transform = CGAffineTransformRotate(transform, (CGFloat) (3.0 * M_PI / 2.0));
 			break;
 		case UIImageOrientationLeft:
 			boundHeight = bounds.size.height;
 			bounds.size.height = bounds.size.width;
 			bounds.size.width = boundHeight;
-			transform = CGAffineTransformMakeTranslation(0.0, imageSize.width);
-			transform = CGAffineTransformRotate(transform, 3.0 * M_PI / 2.0);
+			transform = CGAffineTransformMakeTranslation((CGFloat) 0.0, imageSize.width);
+			transform = CGAffineTransformRotate(transform, (CGFloat) (3.0 * M_PI / 2.0));
 			break;
 		case UIImageOrientationRightMirrored:
 			boundHeight = bounds.size.height;
 			bounds.size.height = bounds.size.width;
 			bounds.size.width = boundHeight;
-			transform = CGAffineTransformMakeScale(-1.0, 1.0);
-			transform = CGAffineTransformRotate(transform, M_PI / 2.0);
+			transform = CGAffineTransformMakeScale((CGFloat) -1.0, (CGFloat) 1.0);
+			transform = CGAffineTransformRotate(transform, (CGFloat) (M_PI /  2.0));
 			break;
 		case UIImageOrientationRight:
 			boundHeight = bounds.size.height;
 			bounds.size.height = bounds.size.width;
 			bounds.size.width = boundHeight;
-			transform = CGAffineTransformMakeTranslation(imageSize.height, 0.0);
-			transform = CGAffineTransformRotate(transform, M_PI / 2.0);
+			transform = CGAffineTransformMakeTranslation(imageSize.height, (CGFloat) 0.0);
+			transform = CGAffineTransformRotate(transform, (CGFloat) (M_PI / 2.0));
 			break;
 		default:
 			[NSException raise:NSInternalInconsistencyException format:@"Invalid image orientation"];
